Add Webpack to Winamp App
