+++ projects/intro/winamp/package.json
@@ -14,6 +14,7 @@
   "devDependencies": {
     "@babel/core": "^7.21.8",
     "@babel/preset-env": "^7.21.5",
+    "@babel/preset-react": "^7.22.0",
     "babel-loader": "^9.1.2",
     "css-loader": "^6.7.4",
     "html-webpack-plugin": "^5.5.1",
@@ -23,5 +24,10 @@
     "webpack-cli": "^5.1.1",
     "webpack-dev-server": "^4.15.0",
     "webpack-merge": "^5.9.0"
+  },
+  "dependencies": {
+    "dotenv": "^16.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
