+++ projects/intro/winamp/src/components/Controls.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export function Controls({ onShuffle }) {
+  return (
+    <div className="control">
+      <button onClick={onShuffle}>Shuffle</button>
+    </div>
+  );
+}
