+++ projects/intro/winamp/src/components/Song.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export function Song({ title, author, children }) {
+  return (
+    <span>
+      {title} - {author} {children}
+    </span>
+  );
+}
