+++ projects/intro/winamp/webpack.prod.js
@@ -3,4 +3,8 @@ const common = require('./webpack.common.js');
 
 module.exports = merge(common, {
   mode: 'production',
+  externals: {
+    react: 'react',
+    'react-dom': 'react-dom',
+  },
 });
